[HACK] integrate topology with scx_p2dq
